\name{is.univariate.ts}
\title{Test for Univariate Time Series}
\usage{
is.univariate.ts (obj)
}
\alias{is.univariate.ts}
\arguments{
  \item{obj}{object to be tested.}
}
\description{
  Returns whether \code{obj} is a univariate time series. 
}
\value{
  A logical.
\author{A. Trapletti}
\keyword{ts}

