\name{NelPlo}
\alias{NelPlo}
\alias{cpi}
\alias{ip}
\alias{gnp.nom}
\alias{vel}
\alias{emp}
\alias{int.rate}
\alias{nom.wages}
\alias{gnp.def}
\alias{money.stock}
\alias{gnp.real}
\alias{stock.prices}
\alias{gnp.capita}
\alias{real.wages}
\alias{unemp}
\title{
  Nelson--Plosser Macroeconomic Time Series
}
\description{
  These are the extended Nelson-Plosser Data. 
}
\usage{
data(NelPlo)
}
\format{
  14 macroeconomic time series: \code{cpi}, \code{ip}, \code{gnp.nom},
  \code{vel}, \code{emp}, \code{int.rate}, \code{nom.wages},
  \code{gnp.def}, \code{money.stock}, \code{gnp.real},
  \code{stock.prices}, \code{gnp.capita}, \code{real.wages}, and
  \code{unemp} and the joint series \code{NelPlo}.
}
\details{
  The series are of various lengths but all end in 1988.
  The data set contains the following series:
  consumer price index, industrial production, nominal GNP, velocity,
  employment, interest rate, nominal wages, GNP deflator, money stock,
  real GNP, stock prices (S&P500), GNP per capita, real wages,
  unemployment.
}
\source{
  \bibshow{Nelson+Plosser:1982}
  
  Formerly in the Journal of Business and Economic Statistics data
  archive, currently at
  \url{http://korora.econ.yale.edu/phillips/data/np&enp.dat}.
}
\references{
  \bibshow{Koop+Steel:1994}
}
\keyword{datasets}
