\name{nino}
\alias{nino}
\title{
  Sea Surface Temperature (SST) Nino 3 and Nino 3.4 Indices
}
\usage{
data (nino)
}
\format{
  Two univariate time series \code{nino3} and \code{nino3.4}
  with 598 observations and the joint series \code{nino}.
}
\description{
  These data constitutes of Nino Region 3 and Nino Region 3.4 SST 
  indices in degrees Celsius. The Nino 3 Region is bounded by 90W-150W
  and 5S-5N. The Nino 3.4 Region is bounded by 120W-170W and 5S-5N. 
}
\source{
  Climate Prediction Center:
  \url{http://www.cpc.ncep.noaa.gov/data/indices/}
}
\keyword{datasets}

