\name{tcm}
\alias{tcm}
\title{
  Yields on Treasury Securities
}
\usage{
data (tcm)
}
\format{
  Two univariate time series \code{tcm10y} and \code{tcm1y}.
}
\description{
  Monthly 10 year and 1 year yields on treasury securities at constant, fixed
  maturity. These are constructed by the Treasury Department, based on
  the most actively traded marketable treasury securities.
}
\source{
  U.S. Fed \url{http://www.federalreserve.gov/Releases/H15/data.htm}
}
\keyword{datasets}


