\name{tcmd}
\alias{tcmd}
\alias{tcm1yd}
\alias{tcm3yd}
\alias{tcm5yd}
\alias{tcm10yd}
\title{
  Daily Yields on Treasury Securities
}
\description{
  This data set contains daily 1 year, 3 year, 5 year, and 10 year
  yields on treasury securities at constant, fixed maturity.
}
\usage{
data(tcmd)
}
\format{
  4 univariate time series \code{tcm1yd}, \code{tcm3yd}, \code{tcm5yd},
  and \code{tcm10yd} and the joint series \code{tcmd}.
}
\details{
  The yields at constant fixed maturity have been constructed by the
  Treasury Department, based on the most actively traded marketable
  treasury securities.

  Daily refers to business days, i.e., weekends and holidays are
  eliminated.
}
\source{
  U.S. Fed \url{https://www.federalreserve.gov/Releases/H15/data.htm}
}
\keyword{datasets}
