\name{tseries.internal}
\title{Internal tseries functions}
\alias{ampsurr}
\alias{boot.sample}
\alias{fftsurr}
\description{
  Internal tseries functions
}
\usage{
ampsurr (x)
boot.sample (x, b, type)
fftsurr (x)
}
\details{
  These are not to be called by the user.
}
\author{A. Trapletti}
\keyword{misc}
\keyword{ts}
