\name{USeconomic}
\alias{USeconomic}
\alias{M1}
\alias{GNP}
\alias{rs}
\alias{rl}
\title{
  U.S. Economic Variables
}
\description{
  This is the E.3 example data set of Luetkepohl (1991).
}
\usage{
data(USeconomic)
}
\format{
  4 univariate time series \code{M1}, \code{GNP}, \code{rs}, and
  \code{rl} and the joint series \code{USeconomic} containing the
  logarithm of \code{M1}, the logarithm of \code{GNP}, \code{rs}, and
  \code{rl}.
}
\details{
  It contains seasonally adjusted real U.S. money \code{M1} and GNP in
  1982 Dollars; discount rate on 91-Day treasury bills \code{rs} and
  yield on long-term treasury bonds \code{rl}. 
}
\source{
  Luetkepohl, H. (1991):
  \emph{Introduction to Multiple Time Series Analysis}.
  Springer Verlag, NY, 500--503.
}
\keyword{datasets}
