\name{bev}
\alias{bev}
\title{
  Beveridge Wheat Price Index, 1500--1869.
}
\description{
  Contains the well-known Beveridge Wheat Price Index which gives annual
  price data from 1500 to 1869, averaged over many locations in western
  and central Europe.
}
\usage{
data(bev)
}
\format{
  A univariate time series with 370 observations.  The object is of
  class \code{"ts"}.
}
\details{
  This data provides an example of long memory time series which has the
  appearance of being nonstationary in levels and yet also appears
  overdifferenced.  See, e.g., Baillie (1996).
}
\source{
  Time Series Data Library:
  \url{https://robjhyndman.com/TSDL/}
}
\references{
  R. T. Baillie (1996):
  Long Memory Processes and Fractional Integration in Econometrics.
  \emph{Journal of Econometrics}, \bold{73}, 5--59.
}
\keyword{datasets}
