\name{NelPlo}
\alias{NelPlo}
\title{
  Nelson-Plosser Macroeconomic Time Series
}
\usage{
data (NelPlo)
}
\format{
  14 macroeconomic time series: \code{cpi}, \code{ip}, \code{gnp.nom},
  \code{vel}, \code{emp}, \code{int.rate}, \code{nom.wages},
  \code{gnp.def}, \code{money.stock}, \code{gnp.real},
  \code{stock.prices}, \code{gnp.capita}, \code{real.wages}, and
  \code{unemp}.
}
\description{
  These are the extended Nelson-Plosser Data. The series are of various
  lengths but all end in 1988. The data set contains the following series:
  consumer price index, industrial production, nominal GNP, velocity,
  employment, interest rate, nominal wages, GNP deflator, money stock,
  real GNP, stock prices (S&P500), GNP per capita, real wages, unemployment.
}
\source{
  Journal of Business and Economic Statistics data archive
  \url{http://www.amstat.org/publications/jbes/}
}
\references{
  G. Koop and M. F. J. Steel (1994): A Decision-Theoretic Analysis of
  the Unit-Root Hypothesis using Mixtures of Elliptical
  Models. \emph{Journal of Business and Economic Statistics} \bold{12}, pp. 95-107.
}
\keyword{datasets}


