\name{NelPlo}
\alias{NelPlo}
\alias{cpi}
\alias{ip}
\alias{gnp.nom}
\alias{vel}
\alias{emp}
\alias{int.rate}
\alias{nom.wages}
\alias{gnp.def}
\alias{money.stock}
\alias{gnp.real}
\alias{stock.prices}
\alias{gnp.capita}
\alias{real.wages}
\alias{unemp}
\title{
  Nelson--Plosser Macroeconomic Time Series
}
\description{
  These are the extended Nelson-Plosser Data. 
}
\usage{
data(NelPlo)
}
\format{
  14 macroeconomic time series: \code{cpi}, \code{ip}, \code{gnp.nom},
  \code{vel}, \code{emp}, \code{int.rate}, \code{nom.wages},
  \code{gnp.def}, \code{money.stock}, \code{gnp.real},
  \code{stock.prices}, \code{gnp.capita}, \code{real.wages}, and
  \code{unemp} and the joint series \code{NelPlo}.
}
\details{
  The series are of various lengths but all end in 1988.
  The data set contains the following series:
  consumer price index, industrial production, nominal GNP, velocity,
  employment, interest rate, nominal wages, GNP deflator, money stock,
  real GNP, stock prices (S&P500), GNP per capita, real wages,
  unemployment.
}
\source{
  C. R. Nelson and C. I. Plosser (1982),
  Trends and Random Walks in Macroeconomic Time Series.
  \emph{Journal of Monetary Economics}, \bold{10}, 139--162.
  \doi{10.1016/0304-3932(82)90012-5}.
  
  Formerly in the Journal of Business and Economic Statistics data
  archive, currently at
  \url{http://korora.econ.yale.edu/phillips/data/np&enp.dat}.
}
\references{
  G. Koop and M. F. J. Steel (1994),
  A Decision-Theoretic Analysis of the Unit-Root Hypothesis using
  Mixtures of Elliptical Models.
  \emph{Journal of Business and Economic Statistics}, \bold{12},
  95--107.
  \doi{10.1080/07350015.1994.10509993}.
}
\keyword{datasets}
