\name{tcm}
\alias{tcm}
\title{
  Yields on Treasury Securities
}
\usage{
data (tcm)
}
\format{
  4 univariate time series \code{tcm1y}, \code{tcm3y}, \code{tcm5y}, and
  \code{tcm10y} and the joint series \code{tcm}.
}
\description{
  Monthly 1 year, 3 year, 5 year, and 10 year yields on treasury
  securities at constant, fixed maturity. These are constructed by
  the Treasury Department, based on the most actively traded marketable
  treasury securities.
}
\source{
  U.S. Fed \url{http://www.federalreserve.gov/Releases/H15/data.htm}
}
\keyword{datasets}


