\name{tcm}
\alias{tcm}
\alias{tcm1y}
\alias{tcm3y}
\alias{tcm5y}
\alias{tcm10y}
\title{
  Monthly Yields on Treasury Securities
}
\description{
  This data set contains monthly 1 year, 3 year, 5 year, and 10 year
  yields on treasury securities at constant, fixed maturity.
}
\usage{
data(tcm)
}
\format{
  4 univariate time series \code{tcm1y}, \code{tcm3y}, \code{tcm5y}, and
  \code{tcm10y} and the joint series \code{tcm}.
}
\details{
  The yields at constant fixed maturity have been constructed by the
  Treasury Department, based on the most actively traded marketable
  treasury securities.
}
\source{
  U.S. Fed \url{https://www.federalreserve.gov/Releases/H15/data.htm}
}
\keyword{datasets}
